7x27_ffa: Resize the memory sent to kernel from 214M to 212M
Resize the memory sent by the bootloader to kernel through kernel
cmdline from 214M to 212M. Actual avialable memory for userspace is
213MB which is further reduced by 1MB such that remaining 1MB can be
used along with 1MB of oemsbl for kgsl buffers of 2MB. Currently
kgsl memory is allocated from the memory sent to kernel. This will
save us 1MB memory for the userspace.

Change-Id: I2944087f125890e3ea22288b1daaf7681786cee6